MainBackend/requirements.txt
Django==5.1.7
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
pytz==2024.1
sqlparse==0.4.4